feat: transform import archive to handler
