+++ api/apps/api/src/modules/clone/import/application/import-application.module.ts
@@ -2,11 +2,11 @@ import { Logger, Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { ImportAdaptersModule } from '../adapters/import-adapters.module';
 import { CompleteImportPieceHandler } from './complete-import-piece.handler';
-import { ImportArchive } from './import-archive';
+import { ImportArchiveHandler } from './import-archive.handler';
 
 @Module({
   imports: [CqrsModule, ImportAdaptersModule],
-  providers: [ImportArchive, CompleteImportPieceHandler, Logger],
+  providers: [ImportArchiveHandler, CompleteImportPieceHandler, Logger],
   controllers: [],
   exports: [],
 })
